refactor: [MoveStrategy] 인터페이스 패키지 구조 변경
