+++ src/main/java/racingcar/model/Cars.java
@@ -1,7 +1,5 @@
 package racingcar.model;
 
-import racingcar.interfaces.MoveStrategy;
-
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
