+++ src/test/java/racingcar/CarsTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.interfaces.MoveStrategy;
+import racingcar.model.MoveStrategy;
 import racingcar.model.Cars;
 
 import static org.assertj.core.api.Assertions.assertThat;
